BTL_ATTT/requirements.txt
Flask==2.3.3
pycryptodome==3.19.0
Werkzeug==2.3.7